+++ hazelcastmq-examples/src/main/java/org/mpilone/hazelcastmq/example/core/NodeFailure.java
@@ -5,14 +5,12 @@ import java.util.concurrent.TimeUnit;
 import javax.jms.JMSException;
 
 import org.mpilone.hazelcastmq.core.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.mpilone.hazelcastmq.example.ExampleApp;
 
-import com.hazelcast.config.Config;
-import com.hazelcast.config.JoinConfig;
-import com.hazelcast.config.NetworkConfig;
-import com.hazelcast.core.Hazelcast;
-import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.config.*;
+import com.hazelcast.core.*;
+import com.hazelcast.logging.ILogger;
+import com.hazelcast.logging.Logger;
 
 /**
  * Example of producing to a queue and then having a node fail. There should be
@@ -20,19 +18,21 @@ import com.hazelcast.core.HazelcastInstance;
  * 
  * @author mpilone
  */
-public class NodeFailure {
+public class NodeFailure extends ExampleApp {
 
-  private final Logger log = LoggerFactory.getLogger(getClass());
+  private final ILogger log = Logger.getLogger(getClass());
 
   private int msgCounter = 0;
-  private String destination = "/queue/node.failure.test";
+  private final String destination = "/queue/node.failure.test";
 
   public static void main(String[] args) throws JMSException,
       InterruptedException {
-    new NodeFailure();
+    NodeFailure app = new NodeFailure();
+    app.runExample();
   }
 
-  public NodeFailure() throws JMSException, InterruptedException {
+  @Override
+  protected void start() throws Exception {
     // Create a three node cluster on localhost. We configure 2 backups so in
     // theory every node should have a complete copy of the queue we're using.
     Config config = new Config();
@@ -103,7 +103,7 @@ public class NodeFailure {
       sendKillTwoRestartOneKillOneAndReceive(node1, node2, node3);
     }
     catch (Throwable ex) {
-      log.error("Unexpected exception during run!", ex);
+      log.severe("Unexpected exception during run!", ex);
     }
     finally {
       node1.shutdown();